Enable iconLeft and iconRight props on <Button />
