+++ src/components/Button/index.js
@@ -1,5 +1,5 @@
 /* @flow */
-export { default as Button } from './Button';
+export { default as Button, RawButton } from './Button';
 export { default as Buttons } from './Buttons';
-export { default as IconButton } from './IconButton';
+export { default as LinkButton } from './LinkButton';
 export { default as PulseButton } from './PulseButton';
