+++ src/index.js
@@ -2,7 +2,13 @@
 import { injectGlobal } from 'styled-components';
 import { Font, FONT_URL } from './lib/constants';
 
-import { Button, Buttons, IconButton, PulseButton } from './components/Button';
+import {
+  Button,
+  Buttons,
+  LinkButton,
+  RawButton,
+  PulseButton,
+} from './components/Button';
 import { Icon } from './components/Icon';
 import { Label } from './components/Label';
 import { Segment } from './components/Segment';
@@ -39,7 +45,8 @@ export const generateAndInjectGlobalCSS = () => injectGlobal`
 
 export {
   Button,
-  IconButton,
+  RawButton,
+  LinkButton,
   PulseButton,
   Buttons,
   Label,
